predictive-maintenance-system/cloud/fault-diagnosis-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
kafka-python-ng==2.2.3
torch==2.5.1
numpy==2.1.3
scikit-learn==1.6.0
networkx==3.4.2
pandas==2.2.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
